requirements.txt
appdirs==1.4.0
argcomplete==1.4.1
numpy==1.11.1
Pillow==3.3.0
pyserial==3.1.1
python-escpos==2.1.3
python-firebase==1.2
pyusb==1.0.0
PyYAML==3.11
qrcode==5.3
requests==2.11.0
six==1.10.0
webkit2png==0.8.2
